+++ eveuniverse/tests/test_core.py
@@ -2,11 +2,12 @@ from unittest.mock import Mock, patch
 
 import requests_mock
 from bravado.exception import HTTPInternalServerError
+from requests.exceptions import HTTPError
 
 from django.core.cache import cache
 from django.test import TestCase
 
-from ..core import esitools, eveimageserver, eveskinserver, fuzzwork
+from ..core import esitools, eveimageserver, evemicros, eveskinserver
 from ..utils import NoSocketsTestCase
 from .testdata.esi import EsiClientStub
 
@@ -203,69 +204,191 @@ class TestEveSkinServer(TestCase):
             eveskinserver.type_icon_url(42, size=22)
 
 
+items = {
+    40170698: {
+        "itemID": 40170698,
+        "itemName": "Colelie VI - Asteroid Belt 1",
+        "typeID": 15,
+        "typeName": "Asteroid Belt",
+        "groupID": 9,
+        "groupName": "Asteroid Belt",
+        "solarSystemID": "30002682",
+        "regionID": 10000032,
+        "constellationID": 10000032,
+        "x": "392074567680",
+        "y": "78438850560",
+        "z": "-199546920960",
+        "security": 0.51238882808862296069918329521897248923778533935546875,
+        "distance": 4.69249999999999989341858963598497211933135986328125,
+        "distanceKm": 701983769,
+    },
+    50011472: {
+        "itemID": 50011472,
+        "itemName": "Stargate (Deltole)",
+        "typeID": 3875,
+        "typeName": "Stargate (Gallente System)",
+        "groupID": 10,
+        "groupName": "Stargate",
+        "solarSystemID": "30002682",
+        "regionID": 10000032,
+        "constellationID": 10000032,
+        "x": "390678650880",
+        "y": "78437130240",
+        "z": "-199573463040",
+        "security": 0.51238882808862296069918329521897248923778533935546875,
+        "distance": 4.6958999999999999630517777404747903347015380859375,
+        "distanceKm": 702495020,
+    },
+    40170697: {
+        "itemID": 40170697,
+        "itemName": "Colelie VI",
+        "typeID": 13,
+        "typeName": "Planet (Gas)",
+        "groupID": 7,
+        "groupName": "Planet",
+        "solarSystemID": "30002682",
+        "regionID": 10000032,
+        "constellationID": 10000032,
+        "x": "390691127743",
+        "y": "78438936622",
+        "z": "-199521990101",
+        "security": 0.51238882808862296069918329521897248923778533935546875,
+        "distance": 4.69620000000000015205614545266143977642059326171875,
+        "distanceKm": 702535753,
+    },
+    40170699: {
+        "itemID": 40170699,
+        "itemName": "Colelie VI - Moon 1",
+        "typeID": 14,
+        "typeName": "Moon",
+        "groupID": 8,
+        "groupName": "Moon",
+        "solarSystemID": "30002682",
+        "regionID": 10000032,
+        "constellationID": 10000032,
+        "x": "390796699186",
+        "y": "78460132168",
+        "z": "-199482549699",
+        "security": 0.51238882808862296069918329521897248923778533935546875,
+        "distance": 4.69620000000000015205614545266143977642059326171875,
+        "distanceKm": 702535998,
+    },
+    40170700: {
+        "itemID": 40170700,
+        "itemName": "Colelie VI - Moon 2",
+        "typeID": 14,
+        "typeName": "Moon",
+        "groupID": 8,
+        "groupName": "Moon",
+        "solarSystemID": "30002682",
+        "regionID": 10000032,
+        "constellationID": 10000032,
+        "x": "390287025280",
+        "y": "78357805096",
+        "z": "-199058647578",
+        "security": 0.51238882808862296069918329521897248923778533935546875,
+        "distance": 4.699799999999999755573298898525536060333251953125,
+        "distanceKm": 703071835,
+    },
+    40170701: {
+        "itemID": 40170701,
+        "itemName": "Colelie VI - Moon 3",
+        "typeID": 14,
+        "typeName": "Moon",
+        "groupID": 8,
+        "groupName": "Moon",
+        "solarSystemID": "30002682",
+        "regionID": 10000032,
+        "constellationID": 10000032,
+        "x": "390135687385",
+        "y": "78327421033",
+        "z": "-198566076258",
+        "security": 0.51238882808862296069918329521897248923778533935546875,
+        "distance": 4.70300000000000029132252166164107620716094970703125,
+        "distanceKm": 703551499,
+    },
+}
+
+
+def create_item(item_id):
+    return items[item_id]
+
+
+def create_request(*item_ids, ok=True):
+    return {
+        "ok": ok,
+        "result": [create_item(item_id) for item_id in item_ids],
+    }
+
+
 @requests_mock.Mocker()
-class TestFuzzworkNearestCelestial(TestCase):
+class TestEveMicrosNearestCelestial(TestCase):
     def setUp(self) -> None:
         cache.clear()
 
     def test_should_return_item_from_api(self, requests_mocker):
         # given
-        item = {
-            "itemName": "Colelie VI - Asteroid Belt 1",
-            "typeid": 15,
-            "itemid": 40170698,
-            "distance": 701983768513.2802,
-        }
         requests_mocker.register_uri(
             "GET",
-            url="https://www.fuzzwork.co.uk/api/nearestCelestial.php?x=660502472160&y=-130687672800&z=-813545103840&solarsystemid=30002682",
-            json=item,
+            url="https://www.kalkoken.org/apps/evemicros/eveUniverse.php?nearestCelestials=30002682,660502472160,-130687672800,-813545103840",
+            json=create_request(40170698, 50011472, 40170697),
         )
         # when
-        result = fuzzwork.nearest_celestial(
+        result = evemicros.nearest_celestial(
             x=660502472160, y=-130687672800, z=-813545103840, solar_system_id=30002682
         )
         # then
         self.assertEqual(result.id, 40170698)
         self.assertEqual(result.name, "Colelie VI - Asteroid Belt 1")
         self.assertEqual(result.type_id, 15)
-        self.assertEqual(result.distance, 701983768513.2802)
+        self.assertEqual(result.distance, 701983769)
         self.assertEqual(requests_mocker.call_count, 1)
 
     def test_should_return_item_from_cache(self, requests_mocker):
         # given
-        item = {
-            "itemName": "Colelie VI - Asteroid Belt 1",
-            "typeid": 15,
-            "itemid": 40170698,
-            "distance": 701983768513.2802,
-        }
         requests_mocker.register_uri(
             "GET",
-            url="https://www.fuzzwork.co.uk/api/nearestCelestial.php?x=1&y=2&z=3&solarsystemid=99",
-            json=item,
+            url="https://www.kalkoken.org/apps/evemicros/eveUniverse.php?nearestCelestials=99,1,2,3",
+            json=create_request(40170698, 50011472, 40170697),
         )
-        fuzzwork.nearest_celestial(x=1, y=2, z=3, solar_system_id=99)
+        evemicros.nearest_celestial(x=1, y=2, z=3, solar_system_id=99)
         # when
-        result = fuzzwork.nearest_celestial(x=1, y=2, z=3, solar_system_id=99)
+        result = evemicros.nearest_celestial(x=1, y=2, z=3, solar_system_id=99)
         # then
         self.assertEqual(result.id, 40170698)
         self.assertEqual(requests_mocker.call_count, 1)
 
     def test_should_return_none_if_nothing_found(self, requests_mocker):
         # given
-        item = {
-            "itemName": None,
-            "typeid": 15,
-            "itemid": 40170698,
-            "distance": 701983768513.2802,
-        }
         requests_mocker.register_uri(
             "GET",
-            url="https://www.fuzzwork.co.uk/api/nearestCelestial.php?x=1&y=2&z=3&solarsystemid=30002682",
-            json=item,
+            url="https://www.kalkoken.org/apps/evemicros/eveUniverse.php?nearestCelestials=30002682,1,2,3",
+            json=create_request(),
         )
         # when
-        result = fuzzwork.nearest_celestial(x=1, y=2, z=3, solar_system_id=30002682)
+        result = evemicros.nearest_celestial(x=1, y=2, z=3, solar_system_id=30002682)
         # then
         self.assertIsNone(result)
+
+    def test_should_return_none_if_api_reports_error(self, requests_mocker):
+        # given
+        requests_mocker.register_uri(
+            "GET",
+            url="https://www.kalkoken.org/apps/evemicros/eveUniverse.php?nearestCelestials=30002682,1,2,3",
+            json=create_request(40170698, 50011472, ok=False),
+        )
+        # when
+        result = evemicros.nearest_celestial(x=1, y=2, z=3, solar_system_id=30002682)
+        # then
+        self.assertIsNone(result)
+
+    def test_should_raise_exception_for_http_errors(self, requests_mocker):
+        # given
+        requests_mocker.register_uri(
+            "GET",
+            url="https://www.kalkoken.org/apps/evemicros/eveUniverse.php?nearestCelestials=30002682,1,2,3",
+            status_code=500,
+        )
+        # when
+        with self.assertRaises(HTTPError):
+            evemicros.nearest_celestial(x=1, y=2, z=3, solar_system_id=30002682)