+++ eveuniverse/tests/test_models_3.py
@@ -4,7 +4,7 @@ import requests_mock
 
 from django.test.utils import override_settings
 
-from ..core import fuzzwork
+from ..core import evemicros
 from ..models import (
     EveAsteroidBelt,
     EveCategory,
@@ -909,13 +909,13 @@ class TestEvePlanetWithSections(NoSocketsTestCase):
         self.assertTrue(obj.enabled_sections.moons)
 
 
-@patch(MODELS_PATH + ".fuzzwork")
+@patch(MODELS_PATH + ".evemicros")
 @patch(MANAGERS_PATH + ".esi")
 class TestEveSolarSystemNearestCelestial(NoSocketsTestCase):
     def test_should_return_stargate(self, mock_esi, mock_fuzzwork):
         # given
         mock_esi.client = EsiClientStub()
-        mock_fuzzwork.nearest_celestial.return_value = fuzzwork.EveItem(
+        mock_fuzzwork.nearest_celestial.return_value = evemicros.EveItem(
             id=50016284, name="Stargate (Akidagi)", type_id=16, distance=1000
         )
         enaluri, _ = EveSolarSystem.objects.get_or_create_esi(id=30045339)
@@ -931,7 +931,7 @@ class TestEveSolarSystemNearestCelestial(NoSocketsTestCase):
     def test_should_return_planet(self, mock_esi, mock_fuzzwork):
         # given
         mock_esi.client = EsiClientStub()
-        mock_fuzzwork.nearest_celestial.return_value = fuzzwork.EveItem(
+        mock_fuzzwork.nearest_celestial.return_value = evemicros.EveItem(
             id=40349471, name="Enaluri III", type_id=13, distance=1000
         )
         enaluri, _ = EveSolarSystem.objects.get_or_create_esi(id=30045339)
@@ -947,7 +947,7 @@ class TestEveSolarSystemNearestCelestial(NoSocketsTestCase):
     def test_should_return_station(self, mock_esi, mock_fuzzwork):
         # given
         mock_esi.client = EsiClientStub()
-        mock_fuzzwork.nearest_celestial.return_value = fuzzwork.EveItem(
+        mock_fuzzwork.nearest_celestial.return_value = evemicros.EveItem(
             id=60015068,
             name="Enaluri V - State Protectorate Assembly Plant",
             type_id=1529,
@@ -966,7 +966,7 @@ class TestEveSolarSystemNearestCelestial(NoSocketsTestCase):
     def test_should_return_asteroid_belt(self, mock_esi, mock_fuzzwork):
         # given
         mock_esi.client = EsiClientStub()
-        mock_fuzzwork.nearest_celestial.return_value = fuzzwork.EveItem(
+        mock_fuzzwork.nearest_celestial.return_value = evemicros.EveItem(
             id=40349487, name="Enaluri III - Asteroid Belt 1", type_id=15, distance=1000
         )
         enaluri, _ = EveSolarSystem.objects.get_or_create_esi(id=30045339)
@@ -985,7 +985,7 @@ class TestEveSolarSystemNearestCelestial(NoSocketsTestCase):
     def test_should_return_moon(self, mock_esi, mock_fuzzwork):
         # given
         mock_esi.client = EsiClientStub()
-        mock_fuzzwork.nearest_celestial.return_value = fuzzwork.EveItem(
+        mock_fuzzwork.nearest_celestial.return_value = evemicros.EveItem(
             id=40349472, name="Enaluri III - Moon 1", type_id=14, distance=1000
         )
         enaluri, _ = EveSolarSystem.objects.get_or_create_esi(id=30045339)
@@ -1001,7 +1001,7 @@ class TestEveSolarSystemNearestCelestial(NoSocketsTestCase):
     def test_should_return_none_if_unknown_type(self, mock_esi, mock_fuzzwork):
         # given
         mock_esi.client = EsiClientStub()
-        mock_fuzzwork.nearest_celestial.return_value = fuzzwork.EveItem(
+        mock_fuzzwork.nearest_celestial.return_value = evemicros.EveItem(
             id=99, name="Merlin", type_id=603, distance=1000
         )
         enaluri, _ = EveSolarSystem.objects.get_or_create_esi(id=30045339)